% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMEselection-package.R, R/GMEselection.R
\docType{package}
\name{GMEselection}
\alias{GMEselection}
\alias{GMEselection-package}
\title{A variable selection procedure for regression models based on Generalized Maximum Entropy estimation (A. Golan, J. Judge and D. Miller, Maximum Entropy Econometrics, Wiley, 1996; Chapter 10). The procedure is useful for well- and ill-posed regression models, namely in models exhibiting small sample sizes, collinearity and non-normal errors.}
\usage{
GMEselection(y, x, ssi = TRUE, si = c(-1, 1), intervals = NULL, m = 5,
  es = TRUE, error.csi = c(-1, 1), j = 3)
}
\arguments{
\item{y}{-- vector -- of size k.}

\item{x}{-- matrix of size (n,k) -- n samples of size k; The supports for all unknown parameters should be symmetric and') uniformly distributed around zero.}

\item{ssi}{(resp) -- TRUE/FALSE -- use the same support interval for all the unknown.}

\item{si}{(int1) -- c(float,float) -- support interval, in the form c(-|lower|,upper), common for all unknown parameters.}

\item{intervals}{-- list of lists -- ex. list(c(-l1,s1), c(-l2,s2),...), list of pairs (.,.) constituting a different support for each parameter. Note that, in this model, you must specify the same number of pairs as size of vector y.}

\item{m}{-- integer -- the number of points in each parameter support. Usually the estimation is performed with five points in the parameter supports. Naturally, you can define a higher value.}

\item{es}{(resp1) -- TRUE/FALSE -- error supports using an estimate of the error standard deviation from the OLS residuals.}

\item{error.csi}{(int2) -- c(-e,e) -- the support interval, [-e,e], for the error component.}

\item{j}{-- integer -- number of points in each error support. Usually the estimation is performed with three points in the error supports. Naturally, you can define a higher value.}
}
\value{
\code{b} -- vector -- estimate of the unknown parameters;
\code{nepk} -- float -- normalized entropy for the intercept (if it exists) and for each variable;
\code{nep} -- vector -- normalized entropy for the signal.
}
\description{
A variable selection procedure for regression models based on Generalized Maximum Entropy estimation (A. Golan, J. Judge and D. Miller, Maximum Entropy Econometrics, Wiley, 1996; Chapter 10). The procedure is useful for well- and ill-posed regression models, namely in models exhibiting small sample sizes, collinearity and non-normal errors.

A variable selection procedure for regression models based on Generalized Maximum Entropy estimation (A. Golan, J. Judge and D. Miller, Maximum Entropy Econometrics, Wiley, 1996; Chapter 10). The procedure is useful for well- and ill-posed regression models, namely in models exhibiting small sample sizes, collinearity and non-normal errors.
}

